+++ shenyu-admin/src/main/java/org/apache/shenyu/admin/service/impl/DashboardUserServiceImpl.java
@@ -20,12 +20,14 @@ package org.apache.shenyu.admin.service.impl;
 import com.google.common.collect.Lists;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
+import org.apache.shenyu.admin.config.properties.DashboardProperties;
 import org.apache.shenyu.admin.config.properties.JwtProperties;
 import org.apache.shenyu.admin.config.properties.LdapProperties;
 import org.apache.shenyu.admin.mapper.DashboardUserMapper;
 import org.apache.shenyu.admin.mapper.RoleMapper;
 import org.apache.shenyu.admin.mapper.UserRoleMapper;
 import org.apache.shenyu.admin.model.dto.DashboardUserDTO;
+import org.apache.shenyu.admin.model.dto.DashboardUserModifyPasswordDTO;
 import org.apache.shenyu.admin.model.dto.UserRoleDTO;
 import org.apache.shenyu.admin.model.entity.DashboardUserDO;
 import org.apache.shenyu.admin.model.entity.RoleDO;
@@ -42,9 +44,11 @@ import org.apache.shenyu.admin.service.publish.UserEventPublisher;
 import org.apache.shenyu.admin.transfer.DashboardUserTransfer;
 import org.apache.shenyu.admin.utils.Assert;
 import org.apache.shenyu.admin.utils.JwtUtils;
-import org.apache.shenyu.admin.utils.ListUtil;
+import org.apache.shenyu.common.utils.ListUtil;
+import org.apache.shenyu.admin.utils.SessionUtil;
+import org.apache.shenyu.admin.utils.ShenyuResultMessage;
 import org.apache.shenyu.common.constant.AdminConstants;
-import org.apache.shenyu.common.utils.ShaUtils;
+import org.apache.shenyu.common.utils.DigestUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.ldap.NameNotFoundException;
@@ -84,13 +88,16 @@ public class DashboardUserServiceImpl implements DashboardUserService {
     
     private final UserEventPublisher publisher;
     
+    private final DashboardProperties properties;
+    
     public DashboardUserServiceImpl(final DashboardUserMapper dashboardUserMapper,
                                     final UserRoleMapper userRoleMapper,
                                     final RoleMapper roleMapper,
                                     @Nullable final LdapProperties ldapProperties,
                                     @Nullable final LdapTemplate ldapTemplate,
                                     final JwtProperties jwtProperties,
-                                    final UserEventPublisher publisher) {
+                                    final UserEventPublisher publisher,
+                                    final DashboardProperties properties) {
         this.dashboardUserMapper = dashboardUserMapper;
         this.userRoleMapper = userRoleMapper;
         this.roleMapper = roleMapper;
@@ -98,6 +105,7 @@ public class DashboardUserServiceImpl implements DashboardUserService {
         this.ldapTemplate = ldapTemplate;
         this.jwtProperties = jwtProperties;
         this.publisher = publisher;
+        this.properties = properties;
     }
     
     /**
@@ -114,6 +122,9 @@ public class DashboardUserServiceImpl implements DashboardUserService {
     
     @Override
     public int create(final DashboardUserDTO dashboardUserDTO) {
+        Assert.notBlack(dashboardUserDTO.getPassword(), "password is not null");
+        Assert.notEmpty(dashboardUserDTO.getRoles(), "role is not empty");
+        Assert.isNull(dashboardUserMapper.selectByUserName(dashboardUserDTO.getUserName()), "the user is existed");
         DashboardUserDO dashboardUserDO = DashboardUserDO.buildDashboardUserDO(dashboardUserDTO);
         // create new user
         final int insertCount = dashboardUserMapper.insertSelective(dashboardUserDO);
@@ -126,7 +137,14 @@ public class DashboardUserServiceImpl implements DashboardUserService {
     
     @Override
     public int update(final DashboardUserDTO dashboardUserDTO) {
+        // 【mandatory】This function can only be used by the admin user
+        Assert.isTrue(SessionUtil.isAdmin(), "This function can only be used by the admin(root) user");
         DashboardUserDO dashboardUserDO = DashboardUserDO.buildDashboardUserDO(dashboardUserDTO);
+        if (Objects.equals(dashboardUserDO.getUserName(), SessionUtil.visitorName())) {
+            Assert.isTrue(Boolean.TRUE.equals(dashboardUserDO.getEnabled()), "You cannot disable yourself");
+        } else {
+            Assert.isTrue(!Objects.equals(dashboardUserDO.getId(), SessionUtil.visitor().getUserId()), "Super administrator name is not allowed to be modified");
+        }
         // update old user
         if (CollectionUtils.isNotEmpty(dashboardUserDTO.getRoles())) {
             if (!AdminConstants.ADMIN_NAME.equals(dashboardUserDTO.getUserName())) {
@@ -135,6 +153,9 @@ public class DashboardUserServiceImpl implements DashboardUserService {
             bindUserRole(dashboardUserDTO.getId(), dashboardUserDTO.getRoles());
         }
         final DashboardUserDO before = dashboardUserMapper.selectById(dashboardUserDO.getId());
+        if (StringUtils.isBlank(dashboardUserDO.getPassword())) {
+            dashboardUserDO.setPassword(before.getPassword());
+        }
         final int updateCount = dashboardUserMapper.updateSelective(dashboardUserDO);
         if (updateCount > 0) {
             publisher.onUpdated(dashboardUserDO, before);
@@ -156,6 +177,9 @@ public class DashboardUserServiceImpl implements DashboardUserService {
                 // skip default admin user
                 .filter(u -> !Objects.equals(u.getUserName(), AdminConstants.ADMIN_NAME))
                 .collect(Collectors.toList());
+        if (CollectionUtils.isEmpty(deletedUser)) {
+            return 0;
+        }
         final List<String> deletedIds = ListUtil.map(deletedUser, DashboardUserDO::getId);
         int deleteCount = dashboardUserMapper.deleteByIdList(deletedIds);
         if (deleteCount > 0) {
@@ -249,13 +273,50 @@ public class DashboardUserServiceImpl implements DashboardUserService {
         
         final LoginDashboardUserVO loginDashboardUserVO = LoginDashboardUserVO.buildLoginDashboardUserVO(dashboardUserVO);
         final DashboardUserVO finalDashboardUserVO = dashboardUserVO;
-        return Optional.ofNullable(loginDashboardUserVO).map(loginUser -> {
-            if (Boolean.FALSE.equals(loginUser.getEnabled())) {
-                return loginUser;
-            }
-            return loginUser.setToken(JwtUtils.generateToken(finalDashboardUserVO.getUserName(), finalDashboardUserVO.getPassword(),
-                    jwtProperties.getExpiredSeconds())).setExpiredTime(jwtProperties.getExpiredSeconds());
-        }).orElse(null);
+        return Optional.ofNullable(loginDashboardUserVO)
+                .map(loginUser -> {
+                    if (Boolean.FALSE.equals(loginUser.getEnabled())) {
+                        return loginUser;
+                    }
+                    return loginUser.setToken(JwtUtils.generateToken(finalDashboardUserVO.getUserName(), finalDashboardUserVO.getPassword(),
+                            jwtProperties.getExpiredSeconds())).setExpiredTime(jwtProperties.getExpiredSeconds());
+                })
+                .orElse(null);
+    }
+    
+    /**
+     * modify password.
+     *
+     * @param dashboardUserModifyPasswordDTO {@linkplain DashboardUserModifyPasswordDTO}
+     * @return rows
+     */
+    @Override
+    public int modifyPassword(final DashboardUserModifyPasswordDTO dashboardUserModifyPasswordDTO) {
+        DashboardUserDO before = dashboardUserMapper.selectById(dashboardUserModifyPasswordDTO.getId());
+        Assert.notNull(before, "current user is not found");
+        Assert.isTrue(Boolean.TRUE.equals(before.getEnabled()), "current user is locked");
+        Assert.isTrue(Objects.equals(before.getPassword(), dashboardUserModifyPasswordDTO.getOldPassword()), "old password is error");
+        
+        DashboardUserDO dashboardUserDO = DashboardUserDO.buildDashboardUserDO(dashboardUserModifyPasswordDTO);
+        int updateCount = dashboardUserMapper.updateSelective(dashboardUserDO);
+        if (updateCount > 0) {
+            publisher.onUpdated(dashboardUserDO, before);
+        }
+        return updateCount;
+    }
+    
+    @Override
+    public boolean checkUserPassword(final String userId) {
+        final DashboardUserDO userDO = dashboardUserMapper.selectById(userId);
+        
+        Assert.isTrue(!Objects.equals(userDO.getDateCreated(), userDO.getDateUpdated()), ShenyuResultMessage.PASSWORD_IS_DEFAULT);
+        
+        // The password has not been changed for a long time
+        Assert.isTrue(passwordUsedLongTime(userDO), ShenyuResultMessage.PASSWORD_USED_FOR_LONG_TIME);
+        
+        // Weak password blacklist
+        
+        return true;
     }
     
     private DashboardUserVO loginByLdap(final String userName, final String password) {
@@ -270,7 +331,7 @@ public class DashboardUserServiceImpl implements DashboardUserService {
                     RoleDO role = roleMapper.findByRoleName("default");
                     DashboardUserDTO dashboardUserDTO = DashboardUserDTO.builder()
                             .userName(userName)
-                            .password(ShaUtils.shaEncryption(password))
+                            .password(DigestUtils.sha512Hex(password))
                             .role(1)
                             .roles(Lists.newArrayList(role.getId()))
                             .enabled(true)
@@ -289,7 +350,7 @@ public class DashboardUserServiceImpl implements DashboardUserService {
     }
     
     private DashboardUserVO loginByDatabase(final String userName, final String password) {
-        return findByQuery(userName, ShaUtils.shaEncryption(password));
+        return findByQuery(userName, DigestUtils.sha512Hex(password));
     }
     
     /**
@@ -308,5 +369,9 @@ public class DashboardUserServiceImpl implements DashboardUserService {
                         .roleId(roleId)
                         .build()))
                 .collect(Collectors.toList()));
+    }
+    
+    private boolean passwordUsedLongTime(final DashboardUserDO userDO) {
+        return userDO.getDateUpdated().getTime() >= System.currentTimeMillis() - properties.getSuperAdminPasswordValidDuration();
     }
 }
